add tag

+++ src/main/java/indi/zhzl/myblognext/model/Blog.java
@@ -4,9 +4,9 @@ import java.io.Serializable;
 import java.util.Date;
 
 public class Blog implements Serializable {
-    private Integer id;
+    private Long id;
 
-    private Integer categoryId;
+    private Long categoryId;
 
     private String title;
 
@@ -26,19 +26,19 @@ public class Blog implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getCategoryId() {
+    public Long getCategoryId() {
         return categoryId;
     }
 
-    public void setCategoryId(Integer categoryId) {
+    public void setCategoryId(Long categoryId) {
         this.categoryId = categoryId;
     }
